+++ dev/line_chart.html
@@ -39,7 +39,7 @@ const dataConfig = {
   nullRatio: 0
 }
 
-const dataManager = mapd3.DataManager()
+const dataManager = mapd3.DataGenerator()
   .setConfig(dataConfig)
 const data = dataManager.generateTestDataset()
 // data.series.forEach((d, i) => {
@@ -49,7 +49,7 @@ const data = dataManager.generateTestDataset()
 // line, area, stackedArea, bar, stackedBar or array of [line,bar]
 // const chartType = data.series.map((d, i) => ["line", "bar"][i % 2])
 // const chartType = ["bar", "bar", "line"]
-const chartType = "stackedBar"
+const chartType = "line"
 
 const colors = mapd3.colors.mapdColors
 const styles = data.series.map((d, i) => ["solid", "dashed", "dotted"][i % 3])
