+++ src/charts/observer.js
@@ -1,6 +1,6 @@
 import * as d3 from "./helpers/d3-service"
 
-import {filterByDate, filterByKey} from "./data-manager"
+import {filterByDate, filterByKey} from "./helpers/common"
 
 export default function Observer (_dataManager) {
   const dispatcher = d3.dispatch("brushChange", "dataFilter")
